requirements.txt
boto3
Pillow
pyheif
